Add support to connect to SOAP via unix socket
This change adds bbstraction for the SOAP client to allow it to connect
using different means of transport. The existing http and https
transports are basicaly unchanged. When the scheme of the URI is `file`
it will assume that a unix socket shall be used and use an according
transport to send and receive SOAP requests via read/write on the
specified socket.