Fix ssoLogon capabilities parameter name
For whatever reason, the SOAP parameter to pass client side capabilities
to the server is different in ssoLogon from what it is named in the
normal logon call. This change corrects the name so the client
capabilities are passed correctly for both logon and ssoLogon. This is
especially important because newer servers require certain client
capabilites to be set.